+++ Services/PatientService.cs
@@ -10,36 +10,47 @@ namespace Services
     {
         private readonly IPatientRepository _repository;
 
-        public PatientService(IPatientRepository repository) => _repository = repository;
+        public PatientService(IPatientRepository repository)
+        {
+            _repository = repository;
+        }
 
         public async Task<PatientDTO> CreateAsync(PatientDTO patientDTO,CancellationToken token)
         {
             var patient = await _repository.CreateAsync(PatientMapper.MapToPatient(patientDTO), token);
 
             if (patient is null)
-            {
                 throw new BadRequestException($"The patient could not be created");
-            }
 
             return PatientMapper.MapToPatientDto(patient);
         }
 
         public async Task DeleteAsync(Guid patientId, CancellationToken token)
         {
+            var patient = await _repository.GetByIdAsync(patientId);
+
+            if (patient is null)
+                throw new EntityNotFoundException("Patient not found");
+
             await _repository.DeleteAsync(patientId, token);
         }
 
-        public async Task<List<PatientDTO>> GetAllAsync(CancellationToken cancellationToken = default)
+        public async Task<IEnumerable<PatientDTO>> GetAllAsync(CancellationToken cancellationToken = default)
         {
             return PatientMapper.MapToPatientDto(await _repository.GetAllAsync(cancellationToken));
         }
 
         public async Task<PatientDTO> GetByIdAsync(Guid patientId, CancellationToken cancellationToken = default)
         {
-            return PatientMapper.MapToPatientDto(await _repository.GetByIdAsync(patientId, cancellationToken));
+            var patient = await _repository.GetByIdAsync(patientId);
+
+            if (patient is null)
+                throw new EntityNotFoundException("Patient not found");
+
+            return PatientMapper.MapToPatientDto(patient);
         }
 
-        public async Task<List<PatientDTO>> SearchByNameAsync(string fullName, CancellationToken token)
+        public async Task<IEnumerable<PatientDTO>> SearchByNameAsync(string fullName, CancellationToken token)
         {
             return PatientMapper.MapToPatientDto(await _repository.SearchByNameAsync(fullName, token));
         }
@@ -49,13 +60,11 @@ namespace Services
             var patient = await _repository.GetByIdAsync(patientId, token);
 
             if (patient is null)
-            {
                 throw new EntityNotFoundException("Patient not found");
-            }
 
             await _repository.UpdateAsync(patientId, PatientMapper.MapToPatient(newPatient), token);
 
-            return PatientMapper.MapToPatientDto(patient);
+            return newPatient;
         }
     }
 }
